+++ app/components/Auth/FullWidthButton.js
@@ -1,7 +1,18 @@
 import React from 'react';
 import { Text, View, TouchableOpacity } from 'react-native';
+import EStyleSheet from 'react-native-extended-stylesheet';
 
-import styles from "./styles";
+const styles = EStyleSheet.create({
+  container: {
+    paddingVertical: 25
+  },
+  label: {
+    color: "white",
+    fontWeight: "700",
+    fontSize: 24,
+    textAlign: "center"
+  },
+});
 
 const FullWidthButton = (props) => {
   return (
