--- app/components/Auth/OutlinedInput/index.js
@@ -1,3 +1,0 @@
-import OutlinedInput from "./OutlinedInput";
-
-export default OutlinedInput
