--- app/components/Auth/OutlinedInput/styles.js
@@ -1,17 +1,0 @@
-import EStyleSheet from 'react-native-extended-stylesheet';
-
-const styles = EStyleSheet.create({
-  inputContainer: {
-    borderRadius: 20,
-    borderWidth: 1,
-    borderColor: "rgba(255,255,255, 0.6)",
-    height: 45,
-    marginBottom: 18,
-    paddingHorizontal: 10
-  },
-  input: {
-    color: "white"
-  },
-});
-
-export default styles;
