--- app/components/Auth/UniversityList/index.js
@@ -1,7 +1,0 @@
-import { connect } from 'react-redux'
-
-import UniversityList from "./UniversityList";
-
-export default connect((state) => ({
-  universities: state.firestore.ordered.universities
-}))(UniversityList)
