--- app/components/Explore/CourseList/index.js
@@ -1,7 +1,0 @@
-import { connect } from 'react-redux'
-import { compose } from 'redux';
-import { withFirebase, firebaseConnect } from 'react-redux-firebase'
-
-import CourseList from "./CourseList";
-
-export default CourseList;
