--- app/components/Explore/CourseList/styles.js
@@ -1,21 +1,0 @@
-import EStyleSheet from 'react-native-extended-stylesheet';
-
-const styles = EStyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  title: {
-    color: '$green',
-    marginLeft: -8
-  },
-  button: {
-    backgroundColor: 'white',
-    borderWidth: 2,
-    borderColor: '$green',
-  },
-  icon: {
-    color: '$green'
-  },
-});
-
-export default styles;
