--- app/components/MyGroups/Document/index.js
@@ -1,3 +1,0 @@
-import Document from "./Document";
-
-export default Document;
