--- app/components/MyGroups/Document/styles.js
@@ -1,10 +1,0 @@
-import EStyleSheet from 'react-native-extended-stylesheet';
-
-const styles = EStyleSheet.create({
-  container: {
-    paddingTop: 10,
-    paddingHorizontal: 15,
-  },
-});
-
-export default styles;
