--- app/components/auth/FullWidthButton/index.js
@@ -1,3 +1,0 @@
-import FullWidthButton from './FullWidthButton';
-
-export default FullWidthButton;
