--- app/components/auth/FullWidthButton/styles.js
@@ -1,15 +1,0 @@
-import EStyleSheet from 'react-native-extended-stylesheet';
-
-const styles = EStyleSheet.create({
-  container: {
-    paddingVertical: 25
-  },
-  label: {
-    color: "white",
-    fontWeight: "700",
-    fontSize: 24,
-    textAlign: "center"
-  },
-});
-
-export default styles;
