+++ app/config/navigationService.js
@@ -1,7 +1,7 @@
 // NavigationService creates an object that is connected to the top level
 //  react component. This allows us to use it as a reference and dispatch
 //  navigation commands from any component in our app. 
-import { NavigationActions } from 'react-navigation';
+import { NavigationActions, StackActions } from 'react-navigation';
 
 let _navigator;
 
@@ -18,9 +18,19 @@ function navigate(routeName, params) {
   );
 }
 
+function push(routeName, params) {
+  const pushAction = StackActions.push({
+    routeName,
+    params
+  });
+
+  _navigator.dispatch(pushAction)
+}
+
 // add other navigation functions that you need and export them
 
 export default {
   navigate,
+  push,
   setTopLevelNavigator,
 };
